inlined system functions
